requirements.txt
ratelimit
python-slack-sdk